requirements.txt
flask
flask-mysql-connector
flask-cors
pyyaml
